Just comments on the bottom
I took a good long look at the code, got a better feel for it and what everything is doing.  I like the flow/logic of it.  I just brainstormed and wrote some things to think about and other things to think about implementing.  I didn't want to mess with your code just yet, I'd like to figure out that site you were talking about with the "e-board" so I can test/debug and whatnot or can i compile this in C++ compilers?