Queue Logic
The code is obviously incorrect but its good enough to follow my thought process. Make a class of Stopsign and then objects of that class Stopsign A, B C, D. Link each set of (two) sensors to corresponding stopsign object. This will allow for easier queue and precedence/priority processing.